[MOVE] Move UpdateUserNicknameRequestDto - UpdateUserNicknameRequestDto 파일 이동
related to #5
